pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.10",
    "fastapi[standard]>=0.128.0",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "openai>=2.15.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "sqlacodegen>=3.2.0",
    "sqlalchemy>=2.0.45",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
